Clean up sample app
